feat(resources): nao permite excluir exercicios, atividades e turmas quando tem dependencies

+++ server/routes/entities/activities.js
@@ -1,6 +1,8 @@
 const express = require('express');
+const HttpStatus = require('http-status-codes');
 
 const ActivityRepository = require('../../domain/repositories/activity');
+const ClassRoomRepository = require('../../domain/repositories/classRoom');
 
 const app = express.Router();
 
@@ -29,8 +31,14 @@ app.post('/', async (request, response) => {
 });
 
 app.delete('/:id', async (request, response) => {
-  await ActivityRepository.destroy({ _id: request.params.id });
-  response.json(true);
+  if (await ClassRoomRepository.find({ 'classworks.activity': request.params.id })) {
+    response.status(HttpStatus.NOT_ACCEPTABLE).json({
+      message: 'Não foi possível excluir esta atividade porque a mesma já está sendo utilizada por uma turma.',
+    });
+  } else {
+    await ActivityRepository.destroy({ _id: request.params.id });
+    response.json(true);
+  }
 });
 
 module.exports = app;
