+++ server/routes/entities/classRooms.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const _ = require('lodash');
+const HttpStatus = require('http-status-codes');
 
 const ClassRoomRepository = require('../../domain/repositories/classRoom');
 const StudentRepository = require('../../domain/repositories/student');
@@ -31,8 +31,14 @@ app.post('/', async (request, response) => {
 });
 
 app.delete('/:id', async (request, response) => {
-  await ClassRoomRepository.destroy({ _id: request.params.id });
-  response.json(true);
+  if (await StudentRepository.find({ classRooms: request.params.id })) {
+    response.status(HttpStatus.NOT_ACCEPTABLE).json({
+      message: 'Não foi possível excluir esta turma porque a mesma já possuí alunos cadastrados.',
+    });
+  } else {
+    await ClassRoomRepository.destroy({ _id: request.params.id });
+    response.json(true);
+  }
 });
 
 app.get('/subscribe/:token', async (request, response) => {
